AIDD_Task4_Assignmet/pyproject.toml
[project]
name = "quiz-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.6.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "pypdf",
]

[tool.setuptools]
py-modules = ["app", "agent", "tools", "main"]
